test: add unit + MockMvc tests across all 10 services + fix QA MEDIUM issues
New test coverage:
- 34 new test files with ~370 test methods across all 10 services
- MockMvc controller tests for all REST controllers (payment, transaction,
  auth, ledger, user, authorization, notification, reporting, retry, scheduler)
- Unit tests for 7 services that had zero tests (user, authorization,
  notification, reporting, retry, scheduler + additional ledger audit)

QA MEDIUM fixes:
- payment: FraudDetectionProperties double→BigDecimal for money thresholds
- transaction: log injection sanitization in TransactionController
- authorization: pagination (Pageable) on GET /admin/roles and /admin/permissions

Stale test fixes:
- BalanceServiceTest: Account constructor, findByAccountNumber
- PaymentServiceTest: FraudScreeningResult API, hasSufficientFunds
- RefreshTokenServiceTest: rewritten for new token rotation API
- LedgerServiceTest: createDoubleEntry individual params

+++ services/payment-service/src/main/java/com/fintech/paymentservice/fraud/rules/AmountAnomalyRule.java
@@ -25,7 +25,7 @@ public class AmountAnomalyRule implements FraudRule {
 
     @Override
     public RuleResult evaluate(Payment payment) {
-        double multiplier = properties.getAmount().getAnomalyMultiplier();
+        BigDecimal multiplier = properties.getAmount().getAnomalyMultiplier();
 
         // Only look at last 90 days, limit to 1000 records to prevent OOM
         Instant windowStart = Instant.now().minus(90, ChronoUnit.DAYS);
@@ -45,7 +45,7 @@ public class AmountAnomalyRule implements FraudRule {
             .reduce(BigDecimal.ZERO, BigDecimal::add);
 
         BigDecimal averageAmount = totalAmount.divide(BigDecimal.valueOf(userPayments.size()), 4, java.math.RoundingMode.HALF_UP);
-        BigDecimal threshold = averageAmount.multiply(BigDecimal.valueOf(multiplier));
+        BigDecimal threshold = averageAmount.multiply(multiplier);
 
         if (payment.getAmount().compareTo(threshold) > 0) {
             return RuleResult.triggered(getName(), properties.getAmount().getRiskScore(),
